requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
aiofiles==23.2.1
python-dotenv==1.0.1
openai==1.12.0
ultralytics==8.1.0
pillow==10.2.0
numpy==1.26.0
